fix: avoid stdlib shadowing by renaming wikibee/types.py to _types.py; update imports and mypy overrides
